Add configuration block with default params to Insights (#52)
* Change dummy app config to postgres

Without the config (and [4.2] included in the migration file) it is
impossible to bootstrap the dummy app, which in turn is required to run
tests. This commit fixes that.

* Add configuration block with default params to Insights

With the config block it becomes possible to attach default params, so
dataclips can be scoped and run in a certain context, for example in SaaS
environments. It simplifies usage in a great deal because one doesn't
have to pass the same param over and over again. It also supports dynamic
values evaluated on-the-fly.

I've used separate configure block (instead of `Dataclips::Engine`) to
loosen a little the hard dependency on Rails. The gem might be used in
other envs as well, so it doesn't have to be tied that hard to Rails.